requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
pyyaml==6.0.1
python-dateutil==2.8.2
matplotlib==3.7.2
seaborn==0.12.2
scipy>=1.10.0
